+++ src/router/PrivateRoute/types.ts
@@ -1,0 +1,7 @@
+import { RouteProps } from "react-router";
+
+export interface PrivateRouteProps extends RouteProps {
+  component: any;
+  redirectPath: string;
+  isLoggedIn: boolean;
+}
